Connect reviews to users

+++ tedMovieApp/tedMovieApp/Services/MovieReviewService.cs
@@ -24,34 +24,43 @@ public class MovieReviewService : IMovieReviewService
         return review ??  throw new InvalidOperationException($"Review with id {id} does not exist");
     }
 
-    public async Task<Review> CreateMovieReview(int movieId, string title, string reviewText, int stars)
+    public async Task<Review> CreateMovieReview(int movieId, string userId, string title, string reviewText, int stars)
     {
         var review = new Review
         {
             Title = title,
             ReviewText = reviewText,
             Stars = stars,
-            MovieId = movieId
+            MovieId = movieId,
+            UserId =  userId
         };
 
         await _movieReviewRepository.Add(review);
         return review;
     }
-    public async Task<Review> DeleteMovieReview(int id)
+    public async Task<Review> DeleteMovieReview(int id, string userId, bool isAdmin)
     {
         var review = await _movieReviewRepository.GetReview(id);
         if (review == null)
             throw new InvalidOperationException($"Review with id {id} does not exist");
         
+        // authorization check
+        if (review.UserId != userId && !isAdmin)
+            throw new UnauthorizedAccessException("You are not allowed to delete this review");
+        
         await _movieReviewRepository.Delete(review);
         return review;
     }
 
-    public async Task<Review> UpdateMovieReview(int id, int movieId, string title, string reviewText, int stars)
+    public async Task<Review> UpdateMovieReview(int id, string userId, bool isAdmin, int movieId, string title, string reviewText, int stars)
     {
         var review = await _movieReviewRepository.GetReview(id);
         if (review == null)
             throw new InvalidOperationException($"Review with id {id} does not exist");
+        
+        // authorization check
+        if (review.UserId != userId && !isAdmin)
+            throw new UnauthorizedAccessException("You are not allowed to update this review");
         
         review.Title = title;
         review.ReviewText = reviewText;
